check timer service

+++ user_server/handlers.py
@@ -2,7 +2,7 @@
 import logging
 from common import mq_pb2
 from common.timer import Timer
-from .shared import timer_dispatcher, receiver, timer_service, const, at_exit, redis
+from .shared import timer_dispatcher, receiver, timer_service, const, at_exit, redis, registry
 
 
 @timer_dispatcher.handler('welcome')
@@ -31,9 +31,10 @@ def on_alarm(id, data: mq_pb2.Alarm):
 
 
 def init():
-    timer_service.call_later('notice', const.RPC_USER, 'notice', delay=10)
-    timer_service.call_repeat('welcome', const.RPC_USER, 'welcome', interval=30)
-    at_exit(lambda: timer_service.remove_timer('welcome', const.RPC_USER))
+    if registry.addresses(const.RPC_TIMER):
+        timer_service.call_later('notice', const.RPC_USER, 'notice', delay=10)
+        timer_service.call_repeat('welcome', const.RPC_USER, 'welcome', interval=30)
+        at_exit(lambda: timer_service.remove_timer('welcome', const.RPC_USER))
     timer = Timer(redis)
     alarm = mq_pb2.Alarm()
     tid = timer.new_stream_timer(alarm, interval=5000, loop=True)
